requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
google-cloud-firestore==2.14.0
stripe==7.8.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
email-validator==2.1.0
